backend/requirements_manual.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
google-generativeai==0.8.3
httpx==0.26.0
redis==5.0.1
celery==5.3.6
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
